Revert "Move PackageConfigFields and related function into a separate package. (#5275)"
This reverts commit c0fe6997f441f35549e13853adccea68f14f6766.
