+++ compiler/damlc/daml-compiler/BUILD.bazel
@@ -35,6 +35,7 @@ da_haskell_library(
         "transformers",
         "unordered-containers",
         "utf8-string",
+        "yaml",
         "zip",
         "zip-archive",
     ],
@@ -51,7 +52,6 @@ da_haskell_library(
         "//compiler/damlc/daml-lf-conversion",
         "//compiler/damlc/daml-opts",
         "//compiler/damlc/daml-opts:daml-opts-types",
-        "//compiler/damlc/daml-package-config",
         "//compiler/damlc/daml-preprocessor",
         "//compiler/repl-service/client",
         "//compiler/scenario-service/client",
